+++ Semantic_App/Semantic/Semantic.csproj
@@ -75,11 +75,17 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Main_form_partial_BrushWithCursor.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="Main_form.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Main_form.Designer.cs">
       <DependentUpon>Main_form.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Main_form_partial_zoomAtCursor.cs">
+      <SubType>Form</SubType>
     </Compile>
     <Compile Include="Network_route_settings.cs">
       <SubType>Form</SubType>
